+++ CHANGELOG.rst
@@ -6,14 +6,34 @@ Changelog
 
 This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
+1.3.1 (2021-04-12)
+------------------
+
+**Added**
+
+**Fixed**
+
+* sync will now not happen if a unmerged cookietemple sync PR is still open
+* cookietemple lint workflow now uses current cookietemple version, where changes were introduced
+
+**Dependencies**
+
+**Deprecated**
+
+
+
 1.3.0 (2021-01-17)
 ------------------
 
 **Added**
 
 * cli-python version 2.0.0 with Poetry, Nox, pre-commit (with black, prettier, [...]), working module autogeneration and much more.
+* [ALL TEMPLATES] Removed Changelog.rst files in favor of Release Drafter, a GitHub Action to automatically include descriptions of changes into a new release
 
 **Fixed**
+
+* blacklisted files for sync defined by the user are now correctly picked up
+* blacklisted sync files, that are newly introduced by a PR are now included in the sync PR, they were introduced but not in the following ones
 
 **Dependencies**